refactor: checker 패키지를 matcher로 이름 변경
