+++ src/main/java/blackjack/domain/result/matcher/MatchResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result.checker;
+package blackjack.domain.result.matcher;
 
 public enum MatchResult {
     MATCH,
