+++ src/test/java/blackjack/domain/result/matcher/PlayerLoseMatcherTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result.checker;
+package blackjack.domain.result.matcher;
 
 import blackjack.domain.card.Number;
 import blackjack.domain.participant.Dealer;
